basic-js: fix disconnect_stream parameters
disconnect_stream takes the receive_mid of the consumer stream we're
receiving on, not the producer stream details. Store the receive_mid
along with the requested producer stream details so we can disconnect it
later on.

Currently, don't merge this as roster updates can fire in parallel, so
running an async handler for onRosterUpdate can cause error noise due to
concurrent roster updates triggering requesting the same stream multiple
times, and we don't want error noise. If/when the vpaas-sdk serialises
processing of roster updates, this can be merged.